Create minor_key.py
In this project we were given a list of musical notes: C, D♭, D, E♭, E, F, G♭, G, A♭, A, B♭, B, C from this list we were supposed to create a harmonic scale using a set sequence of [1 Step] [½ Step] [1 Step] [1 Step] [½ Step] [1½ Step] [½ Step] as well as finding out a way to make the scale loop back to the beginning the sequence went out of bounds. for example, a C scale would be [C, D, E♭, F, G, A♭, B, C]. A G scale would be [G A B♭ C D E♭ G♭ G] and etc. this project is one of many that I have done over the last 2 years of high school and college. It shows my basic understanding of the python language and my ability to be creative when solving problems.