Create ai.facialrecognition
this project is one I created in my free time through going on YouTube and learning more about artificial intelligence/machine learning.
in this project, I used python to create a program that uses the camera of a device(laptop cam/webcam) in order to track all faces detected on the camera. Furthermore, I used a machine learned program called face cascade from open-cv that is a model that can accurately identify faces. My job was just to create the tracking square around the face each time it is identified. 
credit for the project goes to Paul Mchortor on Youtube